Consuming the Discount.Grpc service from Basket.API microservice for applying the dicount on the procuts.

+++ src/Services/Basket/Basket.API/Models/ShoppingCart.cs
@@ -4,7 +4,7 @@ public class ShoppingCart
 {
     public string UserName { get; set; } = default!;
     public List<ShoppingCartItem> Items { get; set; } = new();
-    public decimal TotalPrice => Items.Sum(item => item.Price * item.Quantity);
+    public double TotalPrice => Items.Sum(item => item.Price * item.Quantity);
 
     // Constructor to initialize ShoppingCart with a user name
     public ShoppingCart(string userName)
